ThreadsListActivity: bring back comment count

+++ app/src/main/java/in/shick/diode/common/util/Util.java
@@ -180,7 +180,7 @@ public class Util {
         return getTimeAgo((long)utcTimeSeconds, resources);
     }
 
-    public static String showNumComments(int comments) {
+    public static String showNumComments(Long comments) {
         if (comments == 1) {
             return "1 comment";
         } else {
